MAJOR Overhaul
A lot of stuff has just been pushed!
For one - Querying a POST is now a thing! Check out the brand new read me on the params and how to do that.

This is all also now hosted on GitHub Pages! (PHP calls are to https://api.plaguecraft.xyz with the exact files you see in this repo, considering GitHub Pages only serves static content like HTML).